backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
google-auth==2.23.0

# Data Processing
pandas>=2.2.0
numpy==1.26.2
scipy==1.11.4

# OpenOA - Wind farm operational analysis
openoa>=2.2.0

# File Handling
aiofiles==23.2.1
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
